requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
datasets==4.0.0
dill==0.3.8
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
grpcio==1.74.0
hf-xet==1.1.5
huggingface-hub==0.34.1
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
Markdown==3.8.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.8.55
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.0.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytorch-triton==3.4.0+git11ec6354
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.9.0.dev20250726+cu128
torchaudio==2.8.0.dev20250727+cu128
torchvision==0.24.0.dev20250727+cu128
tqdm==4.67.1
transformers==4.54.0
triton==3.3.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
